% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_to_DDM.R
\name{DD_to_DDM}
\alias{DD_to_DDM}
\title{DD_to_DDM}
\usage{
DD_to_DDM(DD_input, axis)
}
\arguments{
\item{DD_input}{FINISH}

\item{axis}{FINISH}
}
\value{
DDM FINISH
}
\description{
DD_to_DDM
}
\examples{
library(sp)
DD_to_DDM(DD_input = 63.5, axis = "vertical")

}
