% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_to_DMS.R
\name{DD_to_DMS}
\alias{DD_to_DMS}
\title{DD_to_DMS}
\usage{
DD_to_DMS(DD_input, axis)
}
\arguments{
\item{DD_input}{FINISH}

\item{axis}{FINISH}
}
\value{
DMS
}
\description{
DD_to_DMS
}
\examples{
library(sp)
DD_to_DMS(DD_input = 63.5, axis = "vertical")

}
