% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_to_UTM.R
\name{DD_to_UTM}
\alias{DD_to_UTM}
\title{DD_to_UTM}
\usage{
DD_to_UTM(x, y, ellipsoid = "WGS84", return = "all")
}
\arguments{
\item{x}{FINISH}

\item{y}{FINISH}

\item{ellipsoid}{FINISH}

\item{return}{FINISH}
}
\value{
UTM FINISH
}
\description{
DD_to_UTM
}
\examples{
library(sp)
DD_to_UTM(x = 173.5, y = 63.5)

}
