% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDM_to_DD.R
\name{DDM_to_DD}
\alias{DDM_to_DD}
\title{DDM_to_DD}
\usage{
DDM_to_DD(DDM_input, axis)
}
\arguments{
\item{DDM_input}{FINISH}

\item{axis}{FINISH}
}
\value{
DD
}
\description{
DDM_to_DD
}
\examples{
library(sp)
DDM_to_DD(DDM_input = "63° 30 N", axis = "vertical")

}
