% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDM_to_DMS.R
\name{DDM_to_DMS}
\alias{DDM_to_DMS}
\title{DDM_to_DMS}
\usage{
DDM_to_DMS(DDM_input, axis)
}
\arguments{
\item{DDM_input}{FINISH}

\item{axis}{FINISH}
}
\value{
DMS FINISH
}
\description{
DDM_to_DMS
}
\examples{
library(sp)
DDM_to_DMS(DDM_input = "63° 30 N", axis = "vertical")

}
