% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDM_to_UTM.R
\name{DDM_to_UTM}
\alias{DDM_to_UTM}
\title{DDM_to_UTM}
\usage{
DDM_to_UTM(x, y, ellipsoid = "WGS84", return = "all")
}
\arguments{
\item{x}{FINISH}

\item{y}{FINISH}

\item{ellipsoid}{FINISH}

\item{return}{FINISH}
}
\value{
UTM FINISH
}
\description{
DDM_to_UTM
}
\examples{
library(sp)
DDM_to_UTM(x = "173° 30 E", y = "63° 30 N")

}
