% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMS_to_DD.R
\name{DMS_to_DD}
\alias{DMS_to_DD}
\title{DMS_to_DD}
\usage{
DMS_to_DD(DMS_input, axis)
}
\arguments{
\item{DMS_input}{FINISH}

\item{axis}{FINISH}
}
\value{
DD FINISH
}
\description{
DMS_to_DD
}
\examples{
library(sp)
DMS_to_DD(DMS_input = "63° 30' 0\" N", axis = "vertical")
DMS_to_DD(DMS_input = "63° 30' 0 N", axis = "vertical")

}
