% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMS_to_DDM.R
\name{DMS_to_DDM}
\alias{DMS_to_DDM}
\title{DMS_to_DDM}
\usage{
DMS_to_DDM(DMS_input, axis)
}
\arguments{
\item{DMS_input}{FINISH}

\item{axis}{FINISH}
}
\value{
DDM FINISH
}
\description{
DMS_to_DDM
}
\examples{
library(sp)
DMS_to_DDM(DMS_input = "63° 30' 0\" N", axis = "vertical")
DMS_to_DDM(DMS_input = "63° 30' 0 N", axis = "vertical")

}
