% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMS_to_UTM.R
\name{DMS_to_UTM}
\alias{DMS_to_UTM}
\title{DMS_to_UTM}
\usage{
DMS_to_UTM(x, y, ellipsoid = "WGS84", return = "all")
}
\arguments{
\item{x}{FINISH}

\item{y}{FINISH}

\item{ellipsoid}{FINISH}

\item{return}{FINISH}
}
\value{
UTM FINISH
}
\description{
DMS_to_UTM
}
\examples{
library(sp)
DMS_to_UTM(x = "173째 30' 0\" E", y = "63째 30' 0\" N")
DMS_to_UTM(x = "173째 30' 0 E", y = "63째 30' 0 N")

}
