% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTM_to_DD.R
\name{UTM_to_DD}
\alias{UTM_to_DD}
\title{UTM_to_DD}
\usage{
UTM_to_DD(x, y, zone, ellipsoid = "WGS84", return = "all")
}
\arguments{
\item{x}{FINISH}

\item{y}{FINISH}

\item{zone}{FINISH}

\item{ellipsoid}{FINISH}

\item{return}{FINISH}
}
\value{
DD FINISH
}
\description{
UTM_to_DD
}
\examples{
library(sp)
UTM_to_DD(x = 624436.9, y = 7043729.8, zone = 59)

}
