% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTM_to_DDM.R
\name{UTM_to_DDM}
\alias{UTM_to_DDM}
\title{UTM_to_DDM}
\usage{
UTM_to_DDM(x, y, zone, ellipsoid = "WGS84", return = "all")
}
\arguments{
\item{x}{}

\item{return}{}
}
\value{
DDM
}
\description{
UTM_to_DDM
}
