% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTM_to_DMS.R
\name{UTM_to_DMS}
\alias{UTM_to_DMS}
\title{UTM_to_DMS}
\usage{
UTM_to_DMS(x, y, zone, ellipsoid = "WGS84", return = "all")
}
\arguments{
\item{x}{}

\item{return}{}
}
\value{
DMS
}
\description{
UTM_to_DMS
}
