% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTM_to_DDM.R
\name{UTM_to_DDM}
\alias{UTM_to_DDM}
\title{UTM_to_DDM}
\usage{
UTM_to_DDM(x, y, zone, ellipsoid = "WGS84", return = "all")
}
\arguments{
\item{x}{FINISH}

\item{y}{FINISH}

\item{zone}{FINISH}

\item{ellipsoid}{FINISH}

\item{return}{FINISH}
}
\value{
DDM FINISH
}
\description{
UTM_to_DDM
}
\examples{
library(sp)
UTM_to_DDM(x = 624436.9, y = 7043729.8, zone = 59)

}
